+++ app/src/configs/tableConfigs/recordingsTableConfig.ts
@@ -1,8 +1,3 @@
-import RecordingsActionCell from "../../components/recordings/partials/RecordingsActionCell";
-import RecordingsNameCell from "../../components/recordings/partials/RecordingsNameCell";
-import RecordingsStatusCell from "../../components/recordings/partials/RecordingsStatusCell";
-import RecordingsUpdateCell from "../../components/recordings/partials/RecordingsUpdateCell";
-
 /**
  * Config that contains the columns and further information regarding recordings. These are the information that never or hardly changes.
  * That's why it is hard coded here and not fetched from server.
@@ -44,15 +39,4 @@ export const recordingsTableConfig = {
 	resource: "recordings",
 	category: "recordings",
 	multiSelect: false,
-};
-
-/**
- * This map contains the mapping between the template strings above and the corresponding react component.
- * This helps to render different templates of cells more dynamically
- */
-export const recordingsTemplateMap = {
-	RecordingsActionCell: RecordingsActionCell,
-	RecordingsNameCell: RecordingsNameCell,
-	RecordingsStatusCell: RecordingsStatusCell,
-	RecordingsUpdateCell: RecordingsUpdateCell,
 };
